Hw5

+++ core/model/src/main/java/com/example/model/AccountState.kt
@@ -1,8 +1,10 @@
 package com.example.model
 
+import java.math.BigDecimal
+
 data class AccountState(
     val id: Int,
     val name: String,
-    val balance: Float,
+    val balance: BigDecimal,
     val currency: String
 )
