+++ core/model/src/main/java/com/example/model/Transaction.kt
@@ -1,10 +1,12 @@
 package com.example.model
 
+import java.math.BigDecimal
+
 data class Transaction(
     val id: Int,
     val account: BankAccount,
     val category: Category,
-    val amount: Float,
+    val amount: BigDecimal,
     val transactionDate: String,
     val comment: String? = null
 )
